refactor: deadline policies namespace
